+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/dto/QuestionPaymentHistory.java
@@ -1,6 +1,7 @@
-package com.eager.questioncloud.core.domain.payment;
+package com.eager.questioncloud.core.domain.payment.dto;
 
 import com.eager.questioncloud.core.domain.coupon.enums.CouponType;
+import com.eager.questioncloud.core.domain.payment.enums.QuestionPaymentStatus;
 import com.eager.questioncloud.core.domain.question.Subject;
 import java.time.LocalDateTime;
 import java.util.List;
