+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/enums/QuestionPaymentStatus.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.payment;
+package com.eager.questioncloud.core.domain.payment.enums;
 
 import lombok.Getter;
