+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/infrastructure/QuestionPaymentEntity.java
@@ -1,5 +1,7 @@
-package com.eager.questioncloud.core.domain.payment;
+package com.eager.questioncloud.core.domain.payment.infrastructure;
 
+import com.eager.questioncloud.core.domain.payment.enums.QuestionPaymentStatus;
+import com.eager.questioncloud.core.domain.payment.model.QuestionPayment;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
