+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/infrastructure/QuestionPaymentOrderRepository.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.payment;
+package com.eager.questioncloud.core.domain.payment.infrastructure;
 
+import com.eager.questioncloud.core.domain.payment.model.QuestionPaymentOrder;
 import java.util.List;
 
 public interface QuestionPaymentOrderRepository {
