+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/payment/model/QuestionPayment.java
@@ -1,7 +1,8 @@
-package com.eager.questioncloud.core.domain.payment;
+package com.eager.questioncloud.core.domain.payment.model;
 
 import com.eager.questioncloud.core.domain.coupon.enums.CouponType;
 import com.eager.questioncloud.core.domain.coupon.model.Coupon;
+import com.eager.questioncloud.core.domain.payment.enums.QuestionPaymentStatus;
 import com.eager.questioncloud.core.domain.question.Question;
 import java.time.LocalDateTime;
 import java.util.List;
